requirements.txt
python-dotenv==1.1.1
python_dateutil==2.9.0.post0
pytz==2024.1
Requests==2.32.4
